requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-modeltranslation==0.17.3
django-rosetta==0.9.6
django-taggit==1.5.1
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
Pillow==8.3.1
polib==1.1.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
urllib3==1.26.6
whitenoise==5.3.0
